Avoid privilige problems in RSO with early init
Initialize the Mutiny Intfrastructure class eagerly because it reads
system properties and does service loading.

Doing this allows us to remove the last custom logic from our
ReactiveStreamsEngine implementation so we can just use the Mutiny one
directly.
