+++ dev/io.openliberty.io.smallrye.reactive.mutiny/bnd.bnd
@@ -23,6 +23,11 @@ Export-Package: \
   io.smallrye.mutiny.*; version="2.5.1", \
   mutiny.zero.flow.adapters.*; version="1.0.0"
 
+Private-Package: \
+  io.openliberty.mutiny.internal
+
+Bundle-Activator: io.openliberty.mutiny.internal.MutinyActivator
+
 WS-TraceGroup: REACTIVESTREAMS
 
 instrument.disabled: true
@@ -30,3 +35,5 @@ instrument.disabled: true
 -buildpath: \
   io.smallrye.reactive:mutiny;version=2.5.1,\
   io.smallrye.reactive:mutiny-zero-flow-adapters;version=1.0.0,\
+  com.ibm.websphere.org.osgi.core;version=latest,\
+  com.ibm.ws.logging.core;version=latest
