+++ dev/io.openliberty.io.smallrye.reactive.mutiny/src/io/openliberty/mutiny/internal/package-info.java
@@ -1,0 +1,13 @@
+/*******************************************************************************
+ * Copyright (c) 2023 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
+@TraceOptions(traceGroup = "REACTIVESTREAMS")
+package io.openliberty.mutiny.internal;
+
+import com.ibm.websphere.ras.annotation.TraceOptions;
