+++ dev/io.openliberty.microprofile.reactive.streams.operators30.internal/test/src/io/openliberty/microprofile/reactive/streams/test/AbstractReactiveUnitTest.java
@@ -16,6 +16,9 @@ import java.util.concurrent.TimeoutException;
 
 import org.eclipse.microprofile.reactive.streams.operators.PublisherBuilder;
 import org.eclipse.microprofile.reactive.streams.operators.ReactiveStreams;
+import org.eclipse.microprofile.reactive.streams.operators.core.ReactiveStreamsEngineResolver;
+import org.eclipse.microprofile.reactive.streams.operators.core.ReactiveStreamsFactoryImpl;
+import org.eclipse.microprofile.reactive.streams.operators.spi.ReactiveStreamsFactoryResolver;
 import org.junit.After;
 import org.junit.Before;
 
@@ -29,13 +32,15 @@ public class AbstractReactiveUnitTest {
     @Before
     public void activateEngine() {
         engine = new TestReactiveStreamsEngine();
-        engine.activate();
+        ReactiveStreamsEngineResolver.setInstance(engine);
+        ReactiveStreamsFactoryResolver.setInstance(new ReactiveStreamsFactoryImpl());
     }
 
     @After
     public void deactivateEngine() {
         if (engine != null) {
-            engine.deactivate();
+            ReactiveStreamsFactoryResolver.setInstance(null);
+            ReactiveStreamsEngineResolver.setInstance(null);
             engine = null;
         }
     }
